Incorporate a ton of external libraries.
These are more-or-less all dependencies of MuPDF, but not all
of them have CMake support, and some are fairly old-style.
The goal is to support build with Visual Studio, as well as
the more ubiquitous built-in support for Makefile based systems.

This version seems to work on most systems except Xcode. That
would seem to be related to the CMake generator, and might be
investigated later.